requirements.txt
Django
Pillow
django-jazzmin
django-extensions
djangorestframework
drf-spectacular
drf-spectacular-sidecar